requirements.txt
# Web API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Vision
opencv-python==4.8.1.78
mediapipe==0.10.8

# Hardware control
gpiozero==2.0
pigpio==1.78

# Configuration
pyyaml==6.0.1

# Utilities
python-multipart==0.0.6
